alx_travel_app/requirement.txt
python3 -m pip freeze > requirements.txt
django-environ
django
djangorestframework
django-cors-headers 
celery
rabbitmq
drf-yasg
